+++ plugins/destination/clickhouse/typeconv/ch/values/extension.go
@@ -2,7 +2,7 @@ package values
 
 import (
 	"github.com/apache/arrow/go/v13/arrow/array"
-	"github.com/cloudquery/plugin-sdk/v2/types"
+	"github.com/cloudquery/plugin-sdk/v3/types"
 	"github.com/google/uuid"
 )
 
@@ -10,7 +10,7 @@ func extensionValue(arr array.ExtensionArray) any {
 	switch arr := arr.(type) {
 	case *types.UUIDArray:
 		return primitiveValue[uuid.UUID](arr)
-	case *types.InetArray, *types.MacArray, *types.JSONArray:
+	case *types.InetArray, *types.MACArray, *types.JSONArray:
 		return valueStrData(arr)
 	default:
 		// we fallback here to string representation