+++ plugins/destination/test/client/client.go
@@ -8,8 +8,8 @@ import (
 
 	"github.com/apache/arrow/go/v13/arrow"
 	"github.com/cloudquery/plugin-pb-go/specs"
-	"github.com/cloudquery/plugin-sdk/v2/plugins/destination"
-	"github.com/cloudquery/plugin-sdk/v2/schema"
+	"github.com/cloudquery/plugin-sdk/v3/plugins/destination"
+	"github.com/cloudquery/plugin-sdk/v3/schema"
 	"github.com/rs/zerolog"
 )
 
@@ -34,16 +34,16 @@ func (*Client) Metrics() destination.Metrics {
 	return destination.Metrics{}
 }
 
-func (*Client) Read(context.Context, *arrow.Schema, string, chan<- arrow.Record) error {
+func (*Client) Read(context.Context, *schema.Table, string, chan<- arrow.Record) error {
 	return nil
 }
 
-func (*Client) Migrate(context.Context, schema.Schemas) error {
+func (*Client) Migrate(context.Context, schema.Tables) error {
 	return nil
 }
 
 //revive:disable We need to range over the channel to clear it, but revive thinks it can be removed
-func (c *Client) Write(_ context.Context, _ schema.Schemas, records <-chan arrow.Record) error {
+func (c *Client) Write(_ context.Context, _ schema.Tables, records <-chan arrow.Record) error {
 	if c.spec.ErrorOnWrite {
 		return errors.New("error_on_write is true")
 	}
@@ -53,7 +53,7 @@ func (c *Client) Write(_ context.Context, _ schema.Schemas, records <-chan arrow
 	return nil
 }
 
-func (c *Client) WriteTableBatch(_ context.Context, _ *arrow.Schema, records []arrow.Record) error {
+func (c *Client) WriteTableBatch(_ context.Context, _ *schema.Table, records []arrow.Record) error {
 	defer func() {
 		for _, record := range records {
 			record.Release()
@@ -69,6 +69,6 @@ func (*Client) Close(context.Context) error {
 	return nil
 }
 
-func (*Client) DeleteStale(context.Context, schema.Schemas, string, time.Time) error {
+func (*Client) DeleteStale(context.Context, schema.Tables, string, time.Time) error {
 	return nil
 }