+++ plugins/source/datadog/resources/services/dashboards/dashboards.go
@@ -1,6 +1,8 @@
 package dashboards
 
 import (
+	"context"
+
 	"github.com/DataDog/datadog-api-client-go/v2/api/datadogV1"
 	"github.com/cloudquery/cloudquery/plugins/source/datadog/client"
 	"github.com/cloudquery/plugin-sdk/v2/schema"
@@ -32,4 +34,15 @@ func Dashboards() *schema.Table {
 			},
 		},
 	}
+}
+
+func fetchDashboards(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
+	c := meta.(*client.Client)
+	ctx = c.BuildContextV1(ctx)
+	resp, _, err := c.DDServices.DashboardsAPI.ListDashboards(ctx)
+	if err != nil {
+		return err
+	}
+	res <- resp.GetDashboards()
+	return nil
 }