+++ plugins/source/datadog/resources/services/monitors/monitors.go
@@ -1,6 +1,8 @@
 package monitors
 
 import (
+	"context"
+
 	"github.com/DataDog/datadog-api-client-go/v2/api/datadogV1"
 	"github.com/cloudquery/cloudquery/plugins/source/datadog/client"
 	"github.com/cloudquery/plugin-sdk/v2/schema"
@@ -46,4 +48,15 @@ func Monitors() *schema.Table {
 			MonitorDowntimes(),
 		},
 	}
+}
+
+func fetchMonitors(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
+	c := meta.(*client.Client)
+	ctx = c.BuildContextV1(ctx)
+	resp, _, err := c.DDServices.MonitorsAPI.ListMonitors(ctx)
+	if err != nil {
+		return err
+	}
+	res <- resp
+	return nil
 }